refactor(wpdaemon): split connection.rs into submodules (close #43)
